Compiler api (#2244)
