+++ qiskit/assembler/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2019, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""Helper module for Qiskit assembler.
+
+"""
+
+from .assemble_circuits import assemble_circuits
+from .assemble_schedules import assemble_schedules
+from .disassemble import disassemble
+from .run_config import RunConfig
