+++ qiskit/assembler/disassemble.py
