+++ qiskit/assembler/models.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2019, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""Models for RunConfig."""
+
+from qiskit.validation import BaseSchema
+
+
+class RunConfigSchema(BaseSchema):
+    """Schema for RunConfig."""
+
+    # Required properties.
+    # None
+
+    # Optional properties.
