+++ qiskit/compiler/transpile.py
@@ -10,9 +10,8 @@ import warnings
 
 from qiskit.transpiler import Layout, CouplingMap
 from qiskit.tools.parallel import parallel_map
-from qiskit.transpiler.preset_passmanagers import (default_pass_manager_simulator,
-                                                   default_pass_manager)
-from qiskit.compiler.transpile_config import TranspileConfig
+from qiskit.transpiler.transpile_config import TranspileConfig
+from qiskit.transpiler.transpile_circuit import transpile_circuit
 from qiskit.pulse import Schedule
 
 
@@ -147,9 +146,7 @@ def transpile(circuits,
     return circuits
 
 
-# FIXME: (Ali) This function should really be public and have the
-# signature (circuit, transpile_config). It is currently written like this
-# because our parallel tool can only parallelize functions over one variable.
+# FIXME: This is a helper function because of parallel tools.
 def _transpile_circuit(circuit_config_tuple):
     """Select a PassManager and run a single circuit through it.
 
@@ -163,19 +160,7 @@ def _transpile_circuit(circuit_config_tuple):
     """
     circuit, transpile_config = circuit_config_tuple
 
-    # if the pass manager is not already selected, choose an appropriate one.
-    if transpile_config.pass_manager:
-        pass_manager = transpile_config.pass_manager
-
-    elif transpile_config.coupling_map:
-        pass_manager = default_pass_manager(transpile_config.basis_gates,
-                                            transpile_config.coupling_map,
-                                            transpile_config.initial_layout,
-                                            transpile_config.seed_transpiler)
-    else:
-        pass_manager = default_pass_manager_simulator(transpile_config.basis_gates)
-
-    return pass_manager.run(circuit)
+    return transpile_circuit(circuit, transpile_config)
 
 
 def _parse_transpile_args(circuits, backend,
